# spec/api.openapi.yaml
openapi: 3.1.0
info:
  title: Call Campaign Service API
  version: 0.1.0
  description: |
    HTTP-контракт сервиса управления обзвоном (тестовое задание).

    Контракт запуска: `docker compose up -d` из корня → API на `http://localhost:8080`,
    `GET /healthz` → `200`, PostgreSQL на `localhost:54329`.

    ## Аутентификация и роли
    - JWT HS256, claims: `sub`, `org_id`, `role`
      (`worker` — сервисные воркеры, `authenticated` — пользователи веб-приложения).
      Принимается либо заголовок `Authorization: Bearer <jwt>` (воркеры / API-клиенты),
      либо HttpOnly cookie `dev_token` (веб-UI после `POST /dev/token`).
      Стенд подписывает токены сам нашим `JWT_SECRET`; выдача токенов — за пределами сервиса.
    - Вебхуки провайдера: HMAC-SHA256 от сырого тела запроса с ключом `WEBHOOK_SECRET`,
      заголовок `X-Signature: sha256=<hex>`.
    - Изоляция организаций — PostgreSQL RLS: чужой ресурс отвечает `404`, а не `403`.
    - Матрица ролей (дублируется в `x-roles` каждой операции): `claim_next_contact` — обе
      роли; `provider-link`/`abort` — только `worker`; `GET /api/call_attempts*` (list/detail/stream)
      — `authenticated`; `/api/analyses*` — обе роли.
      Недопущенная роль → `403` `forbidden` (роль — не изоляция, поэтому не `404`).

    ## Семантика ошибок
    - Все ошибки — `Error` с машиночитаемым `code` (enum `ErrorCode`) и человекочитаемым
      `detail`. Клиенты и стенд опираются на `code`, а не на текст.
...

    ## Идемпотентность и ретраи
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/call_attempts:
    get:
      tags: [call-attempts]
      operationId: listCallAttempts
      x-roles: [authenticated]
      summary: Список попыток звонка организации
      description: |
        Курсорная пагинация по `(created_at DESC, id DESC)`. Возвращает только
        попытки org из JWT (RLS). Расширение сверх ТЗ для веб-UI.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          required: false
          description: Непрозрачный курсор из `next_cursor` предыдущего ответа
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Страница списка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallAttemptListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/call_attempts/stream:
    get:
      tags: [call-attempts]
      operationId: streamCallAttempts
      x-roles: [authenticated]
      summary: SSE-лента изменений статусов попыток org
      description: |
        Live-обновления для списка/деталей. События:

        - `attempt` — изменился статус попытки; data = `CallAttemptListItem`
        - `crm` — изменилась доставка в CRM; data = `{ "attempt_id", "crm" }`
        - `analysis` — изменился статус LLM-разбора; data =
          `{ "attempt_id", "analysis_id", "status" }`

        Механика (как у `/api/analyses/:id/stream`):
        - первая строка — `retry: 3000`;
        - heartbeat `: ping` каждые 15 с молчания;
        - notify — только wake-up; payload берётся повторным SELECT под RLS;
        - клиент — fetch-based SSE с `Authorization` или cookie `dev_token`
          (не `EventSource`).

        Расширение сверх ТЗ для веб-UI.
      responses:
        '200':
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/call_attempts/{attempt_id}:
    get:
      tags: [call-attempts]
      operationId: getCallAttempt
      x-roles: [authenticated]
      summary: Детали попытки со статусной историей, LLM и CRM
      description: |
        Таймлайн `status_history`: синтетическая точка `queued` при claim
        (`source=claim`) + применённые webhook-события (`source=webhook`,
        типы ТЗ → статусы) + точка `abort` при терминировании без вебхуков
        (`source=abort`). Плюс связанные `analyses` и состояние `crm` outbox
        (null, пока терминальный статус не поставил запись в outbox).

        Чужая org → `404`. Расширение сверх ТЗ для веб-UI.
      parameters:
        - $ref: '#/components/parameters/AttemptId'
      responses:
        '200':
          description: Детали попытки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallAttemptDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /dev/token:
    post:
      tags: [dev]
      operationId: mintDevToken
      summary: Выпустить JWT (только разработка)
      description: |
        Доступен только при `DEV_TOKEN_ENABLED=true`, иначе `404`.
        Ответ содержит JWT в теле и выставляет HttpOnly cookie `dev_token`
        (`SameSite=Lax`, `Path=/`; `Secure` — при `COOKIE_SECURE=true`).
        При проверке стенд подписывает токены сам — эндпоинт не участвует в контракте стенда.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevTokenRequest'
      responses:
        '200':
          description: Подписанный JWT (+ Set-Cookie `dev_token`)
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HttpOnly cookie `dev_token=<jwt>`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevTokenResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /dev/logout:
    post:
      tags: [dev]
      operationId: logoutDevToken
      summary: Снять cookie `dev_token` (только разработка)
      description: Доступен только при `DEV_TOKEN_ENABLED=true`, иначе `404`.
      security: []
      responses:
        '200':
          description: Cookie удалена
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                    const: true
        '404':
          $ref: '#/components/responses/NotFound'
  /dev/session:
    get:
      tags: [dev]
      operationId: getDevSession
      summary: Статус сессии по cookie/Bearer (только разработка)
      description: |
        Доступен только при `DEV_TOKEN_ENABLED=true`, иначе `404`.
        Не бросает `401` при отсутствии токена — возвращает `authenticated: false`.
      security: []
      responses:
        '200':
          description: Состояние сессии
          content:
            application/json:
              schema:
                type: object
                required: [authenticated]
                properties:
                  authenticated:
                    type: boolean
                  sub:
                    type: string
                    nullable: true
                  org_id:
                    type: string
                    format: uuid
                    nullable: true
                  role:
                    type: string
                    enum: [worker, authenticated]
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT, claims `sub`, `org_id`, `role`.
        Альтернатива — cookie `dev_token` (см. cookieAuth).
    cookieAuth:
      type: apiKey
      in: cookie
      name: dev_token
      description: HttpOnly JWT cookie, выставляется `POST /dev/token` для веб-UI
    webhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: "`sha256=` + hex(HMAC-SHA256(raw body, WEBHOOK_SECRET))"
  parameters:
    AttemptId:
      name: attempt_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    Unauthorized:
      description: '`code`: `unauthorized` (нет/невалидный JWT) или `token_expired`'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Роль не допущена к операции (`code: forbidden`)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'Не найдено или принадлежит другой организации (`code: not_found`)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 'Тело запроса не соответствует схеме (`code: validation_error`)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Uuid:
      type: string
      format: uuid
    ErrorCode:
      type: string
      description: Машиночитаемый код ошибки; опираться на него, а не на `detail`
      enum:
        - unauthorized
        - token_expired
        - invalid_signature
        - forbidden
        - not_found
        - already_linked
        - not_queued
        - no_transcript
        - attempt_not_completed
        - analysis_terminal
        - validation_error
    Error:
      type: object
      required: [code, detail]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
              description: массив ошибок FastAPI при code=validation_error
...
    CallAttemptStatus:
      type: string
      enum: [queued, dialing, in_progress, completed, failed, no_answer]
      description: Статусы попытки по маппингу типов вебхуков ТЗ
    CallAttemptListItem:
      type: object
      required: [id, status, phone, campaign_name, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        status:
          $ref: '#/components/schemas/CallAttemptStatus'
        phone:
          type: string
          description: contacts.phone_e164
        campaign_name:
          type: string
        started_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        ended_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        created_at:
          type: string
          format: date-time
    CallAttemptListResponse:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CallAttemptListItem'
        next_cursor:
          anyOf:
            - type: string
            - type: 'null'
          description: Курсор следующей страницы; null если страница последняя
    StatusHistorySource:
      type: string
      enum: [claim, webhook, abort]
      description: |
        `claim` — создание попытки; `webhook` — применённое событие провайдера;
        `abort` — `POST .../abort` без вебхуков.
    StatusHistoryItem:
      type: object
      required: [at, status, source]
      additionalProperties: false
      properties:
        at:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/CallAttemptStatus'
        source:
          $ref: '#/components/schemas/StatusHistorySource'
        event_type:
          anyOf:
            - type: string
            - type: 'null'
          description: Тип вебхука (`call.dialing` …); null для claim/abort
        sequence:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: sequence вебхука; null для claim/abort
    CrmDeliveryState:
      type: string
      enum: [pending, retrying, delivered]
      description: |
        `pending` — в outbox, ещё не было неуспешных попыток;
        `retrying` — были ошибки, ждём next_attempt_at;
        `delivered` — delivered_at заполнен.
    CrmDelivery:
      type: object
      required: [state, attempts]
      additionalProperties: false
      properties:
        state:
          $ref: '#/components/schemas/CrmDeliveryState'
        attempts:
          type: integer
          minimum: 0
          description: Число неуспешных/выполненных попыток доставки из outbox
        delivered_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        last_error:
          anyOf:
            - type: string
            - type: 'null'
        next_attempt_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
    CallAttemptContact:
      type: object
      required: [phone, timezone]
      additionalProperties: false
      properties:
        phone:
          type: string
        timezone:
          type: string
          description: IANA timezone контакта
    AnalysisSummary:
      type: object
      required: [id, status, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        status:
          $ref: '#/components/schemas/AnalysisStatus'
        result:
          anyOf:
            - $ref: '../packages/shared/analysis-result.schema.json'
            - type: 'null'
        partial:
          anyOf:
            - $ref: '#/components/schemas/PartialAnalysisResult'
            - type: 'null'
        error:
          anyOf:
            - type: string
            - type: 'null'
        created_at:
          type: string
          format: date-time
    CallAttemptDetail:
      type: object
      required:
        - id
        - status
        - campaign_name
        - contact
        - status_history
        - analyses
        - created_at
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        status:
          $ref: '#/components/schemas/CallAttemptStatus'
        provider_call_id:
          anyOf:
            - type: string
            - type: 'null'
        campaign_name:
          type: string
        contact:
          $ref: '#/components/schemas/CallAttemptContact'
        started_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        ended_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        outcome:
          anyOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        transcript:
          anyOf:
            - type: string
            - type: 'null'
        status_history:
          type: array
          items:
            $ref: '#/components/schemas/StatusHistoryItem'
        analyses:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisSummary'
        crm:
          anyOf:
            - $ref: '#/components/schemas/CrmDelivery'
            - type: 'null'
          description: null если терминальный триггер ещё не создал outbox-запись
        created_at:
          type: string
          format: date-time
...
    AnalysisStatus:
      type: string
      enum: [queued, streaming, done, error, cancelled]
      description: Серверный статус; `reconnecting`/`partial` из ТЗ 7.4 — клиентские производные
    PartialAnalysisResult:
      type: object
      description: >-
        Накопленный частичный результат: та же форма, что у схемы 7.3
        (`packages/shared/analysis-result.schema.json`), но все поля необязательны.
        Правило склейки: дельты чанков одного поля конкатенируются как текст;
        строковые поля — это сам текст, а `objections`/`lead_score`/`confidence`
        стримятся как JSON-представление значения. Поле появляется в `partial`,
        когда его накопленный текст успешно парсится.
      additionalProperties: false
      properties:
        summary:
          type: string
        objections:
          type: array
          items:
            type: string
        next_step:
          type: string
          enum: [call_back, send_proposal, disqualify, escalate]
        lead_score:
          type: integer
          minimum: 0
          maximum: 100
        confidence:
          type: number
          minimum: 0
          maximum: 1
...
    DevTokenRequest:
      type: object
      required: [sub, org_id, role]
      additionalProperties: false
      properties:
        sub:
          type: string
          minLength: 1
        org_id:
          $ref: '#/components/schemas/Uuid'
        role:
          type: string
          enum: [worker, authenticated]
        expires_in:
          type: integer
          minimum: 60
          default: 86400
          description: TTL токена в секундах
    DevTokenResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string